Cleanup the binding of Git into the URL space
When I removed the /p/ directory from the Git URL access I forgot
Guice can actually setup the request's path info for JGit if we
use serveRegex() with an HttpServlet. This vastly cleans up the
code and makes it a little easier to understand the URL mapping.

Switch back to extending GitServlet, and configure Guice to only
trigger on the 3 major Git suffix patterns. If /p is present at
the start of the URL, Guice will omit it from the path info before
passing on the request to JGit.

Because of the legacy redirection support in UrlModule (to point
web browsers from /p/foo to /q/status:open+project:foo,n,z) we must
bind GitOverHttpModule ahead of WebModule (which binds UrlModule).

Change-Id: I238ef64d9b2255f839b3c9c2f1038c5dbe2ff978

Former-commit-id: ac59f1eeda6b43a17f5256b7cbaf589b93a603dc