Bind the project cache clock thread in its own object
This allows me to better control when this thread gets created in a
server. Right now its created 1:1 with the ProjectCacheImpl, which
works fine for servers that have only one site loaded in the JVM.
Longer term we should be supporting virtual hosting, where different
Gerrit Code Review sites are able to be managed from a single JVM,
and a single unified cache footprint (that still shards data uniquely
for each site).

Change-Id: I3e9b269aac7071a2f26a4289b4b524cb468592c7

Former-commit-id: a595eacf30bee966f2c6a891c1dd7a62cd5c5501