Improve 'Push Merge Commit' access right documentation
In the documentation the 'Push Merge Commit' access right was named
'Push Merge'. Correct the name to be consistent with the name that is
displayed in the Gerrit WebUI.

In addition remove the link to the 'Push' access right from the
documentation of the error 'you are not allowed to upload merges'.
When the user gets this error he has already the 'Push' access right
assigned and only 'Push Merge Commit' is missing. If 'Push' would be
missing the user would get an 'Upload Denied' error. Also the
'Push Merge Commit' documentation explains that it requires 'Push'.
This means just having the link to 'Push Merge Commit' should be
sufficiant for the 'you are not allowed to upload merges' error
description.

Change-Id: Idfec58e52ba77c967abb97deb6cc90f42c3395d9
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: aef5d6e001922f1bbeb73fd76e0ab92d8ba3a289