Fix deleting/publishing of draft patch sets for non draft changes
Change I9ef00d6382c60ce28c1c002fc2c4733ac129e716 made it impossible
to delete or publish a draft patch set if the change is not a draft.
It is a valid use case to upload a draft patch set for a published
change and it must be possible to delete or publish this draft patch
set.

Change-Id: I2048582148b37e825bf28956fe665fda428956e9
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 0f9e74948347a4be1f3048625a27e498e5323603