Users who starred a change should receive all the emails about a change.
The "emailOnlyAuthors" feature silences emails about actions performed by
users in some groups. The emails for any actions in these groups are sent
only to change authors. With this commit, if a user has starred a change,
it would override this "silencing" and send them an email even if they are
not an author of that change. This provides users with an easy way to opt
in again on specific change emails.

Change-Id: Id1afb7068fece58a9336cddfbc46f9c4bbad98f2
Bug: issue 904

Former-commit-id: a8c4bd115841cb89ab083e027c18359cfdc03101