Updated to gwtorm 1.4.
Refactored the GWT client code to include only what the gerrit GWT
client. The gerrit GWT build was giving build warnings since the
getAsync() signature and implementation used types that were
not available to GWT.

Split the reviewdb into a client and server package.

Organized the imports to ease the migration.

Change-Id: I6cd6125eddcb1dd2b5c5e353edf163544248026b

Former-commit-id: b30883860d446e2b7cd81a36094e8febc6c7e60d