Clarify the behavior of continuingRequest()
It's not exactly true to say that ThreadLocalRequestScopePropagator
doesn't propagate *any* request-specific information, since the
context implementation type typically does contain some request-
specific information, which can in turn be used to derive the
necessary RequestScoped objects in the new thread.

Clarify this behavior in the javadoc, and give an example. This is a
fairly technical description, but the target audience is someone
implementing a new RequestScopePropagator, so they already need to
understand the request scope internals pretty well.

Change-Id: I3326444c76d305fa4b067288ab2a3c7e4ad32997

Former-commit-id: 5447cd7cd7e43b74677067312d28d89fa55f1c9f