Fix development_become_any_account redirects
In a development environment the guessing of the canonicalWebUrl is
currently broken by 172fa525c991dbc04809710edbe4481d27ffc900 [formerly 3e7ef3cc5f41bc829f5e2951b9699079e48d60a5]. Use the
context path as the redirect when logging in a user to work around any
breakage. Using the real canonicalWebUrl isn't as necessary with this
mode of authentication as it is with OpenID, so this minor deviation
from most other code may be acceptable to keep development
environments working as expected.

Change-Id: I591a58032618ce58f4722270fdc60ed880c50fa3

Former-commit-id: 83fe3e88255170cf62d35805f263eea81b547e5e