Fix: Inconsistent behaviour when replicating refs/meta/config
In replication.config, if authGroup is set to be used together
with mirror = true, refs blocked through the authGroup are deleted
from the slave/mirror. The same still (and rightly) applies if
the authGroup is used to block refs/meta/config.

However, if replicatePermission set to false, gerrit will refuse
to clean up refs/meta/config on the slave/mirror.

This fix makes sure that the Gerrit is still allowed to clean
up all refs on the mirror, including refs/meta/config, even when
used in conjunction with replicatePermission = false.

Change-Id: Iea3f00e976d2d3e0bd8cb632530f65d79d9943ba

Former-commit-id: 189282bac993b5c9b5e63ead0bf3aa74eaaa5856