Branch Operator: Support full branch names
The search operator for branches required the provided value to be the
short branch name that is shown in the web interface (without the
'refs/heads/' prefix). Change the branch operator so that it also
supports full branch names as value.

It is intuive that searching with 'branch:master' and searching with
'branch:refs/for/master' deliver the same result. So far
'branch:refs/for/master' was the same as searching with
'refs:refs/heads/refs/heads/master' which is unexpected for most users.

Change-Id: I573eb4d500160843bc28d142488bb8cbfab5a939
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: ad921cdbd09d26ac465b6dfa31c8bd4033de6db9