Create index for submodule subscriptions on site upgrade
Commit 73305d26e91adafbfca8ae6d00d338219dc8dc5d [formerly d15704079c2f820995ee70e7403dfa7ebcb3d791] introduced a new
database index for accessing submodule subscriptions, but the
corresponding database schema Schema_61 did not create the index. As
result the new index was only created if a new site was
initialized but not when an existing site was upgraded. This change
adds a new database schema that takes care to create the missing
index.

Change-Id: Ib4fa1cf184c91935445ce41f66634cd8a5221a45
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: b0d373c272b814ffc4b43552b1d6a58fb8ec875a