Make ls-projects available on HTTP as GET /projects/
The query string takes the same arguments as the command line tool,
so `ls-projects --type CODE --show-branch foo` can be written out as
`GET /projects/?type=CODE&show-branch=foo`.  Output is identical.

Using /projects/ is anonymous and does not require authentication.
/a/projects/ requires HTTP authentication using the same scheme
that is applied to Git operations over HTTP, typically using the
HTTP password that is generated for a user or SSL certificate auth.

Change-Id: I577ae60eba40cfb358fe41ed8f3cf9481662e738

Former-commit-id: a04189872407773e0a9e53ff3a743b80a4f78a97