SuggestService visibility logic updated.
The SuggestService previously used the project to determine if a
suggested user should be returned. The RPC was updated to use the
change to determine if a suggested user can see the change. This
is implemented as a new RPC, deprecating the old one, so the API
does not break unexpectly for direct API users.

This change is moving toward the goal of not using
GroupMembership.getKnownGroups(), so as to support almost any group
system.

Change-Id: I63a8bca720d0cc77b4a9f3820e1d6ebe8966afc4

Former-commit-id: bd564c810c2e2e5f53efbba3e69a00b31e876919