Deprecate /query
Long term the Gerrit project is going to move to a more REST-ful
API. Using /query to scan for changes doesn't make sense in that
world. It also has some issues with output formatting that doesn't
line up with what we want to process in the stock web UI, and does
not match the JSON or JSON_COMPACT formats used by the existing
REST APIs /projects/ and /accounts/self/capabilities.

Deprecate the /query API, leaving it on by default. Define a new
site configuration variable to allow the URL to be disabled if a
server admin wants to promote tools moving to the new API.

Leaving /query on by default is important right now as the new API
will be defined later, primarily because we may want the class name
ChangeQueryServlet for the new implementation.

Change-Id: I86af3f94f66948e6f45da71e9b8631a1c2d1d135

Former-commit-id: 6bd04fd28bec41428520f1e72fd72106259aacbb