REST API /projects/$SUGGEST
The /projects/ URL now accepts a prefix string as part of the URL,
making it usable by the suggestion client UI to limit results.
With this change, the project suggestion service can be removed
and replaced by the REST API.

"GET /projects/platform/" will list all projects that start with
the specified platform/ prefix.

Change-Id: I6f8b9302166f59d03d58a85292cbf052e8419a78

Former-commit-id: 5cd0577828c5ec798c266c9abcf0c1dc59ebcba5