Remove SparseHtmlFile dependency from PatchScript
My hosting environment builds its own custom Gerrit, without
using Maven. In that build process we need to break the dependency
between PatchScript in the server CLASSPATH and SafeHtml, which
isn't available in the server CLASSPATH as its supposed to only be
used in the GWT JavaScript world.

Move the factory that builds the SparseHtmlFile off PatchScript and
into the real client code that needs it, AbstractPatchContentTable
and its two subclasses.

Change-Id: I7375e0962ef53ff83aee2c6abf53bae5eba25286

Former-commit-id: 78f15571e5da3ea347ddc9098f79949c459f097a