Include the ContributorAgreements in the ProjectConfig.
Update the ProjectConfig parsing to support the contributor agreements.
In future changes, the database will be migrated to only use the
ProjectConfig.  One change from the original schema is autoVerify is
now an explicit group to automatically add user into, instead of a
boolean. Also, there is no active field.

Change-Id: Ib5820eb9c976f7ca20599fb64ad610cfa6370d18

Former-commit-id: 1e637400d122aa176ac739cca7e89c0ca9c52ffb