Fix style sheets for expanded context lines
We were adding the diffText css to context lines which were hidden at
first and then expanded.  This caused those lines to have an extra 2
pixel shift to the right.  The change was very subtle but could be noticed
with close inspection.

Also set the vertical alignment of expanded table rows so they match the
valign property of alredy visible rows.

Signed-off-by: Brad Larson <bklarson@gmail.com>
Change-Id: I9e507400e2780cf13a67b4781b3754443957548f

Former-commit-id: 7e5cc7de82ae33165bc39359fb49cb74c011a0dc