Refactored ContributorAgreements to All-Projects project.config.
Existing usages of ContributorAgreements were updated to use
project.config. Only agreements with autoVerify groups are able
to submit "I AGREE" on the new agreement screen. This represents
an add to the group.

All of the existing review DB accesses and types were removed for
AccountAgreement, AccountGroupAgreement, and ContributorAgreement.
Also, a breaking change was made to ChangeHookRunner, the passes the
cla-name instead of the cla-id.

Change-Id: Idebf7faa1e63588f2a7479dd28e00ebeff06b3a9

Former-commit-id: da02c7576d3b7b211a9167875471c46a84661321