/changes/: Display account dashboards
Replace the old JSON-RPC call showing a dashboard to instead be
three queries using the newer /changes/ REST API. This moves all
of the logic into the common ListChanges and QueryProcessor, and
simplifies the server code.

Like the All > Open view, the columns in the table now update
dynamically to reflect the labels actually required to submit
the changes displayed.

Changes in the outgoing and incoming sections are sorted by project,
branch, topic, and then the legacy change number.  This gives a
more predictable sorting when looking at a very busy dashboard,
as users can scan the Project and Branch columns and work down
the page until they locate the change(s) they are looking for. In
particular it helps me when I want to look at changes for git-repo,
or stable changes for gerrit near release time.

Change-Id: Ia9e9fc103561b43fe0188958bfa226fdc21b6992

Former-commit-id: 2a912fba44b88c2650ea2a5a29e69ff24ed6cb2f