Improve the side-by-side viewer table
Move the line number column for the right side to be on the far
right of the table, giving us the layout:

  1 |  foo       |       bar   | 1
  2 |  hello     |       hello | 2

This looks nicer when reading a lot of code, as the line numbers
are less relevant entities than the code itself which is now in
the center of the UI.

Line numbers are still links to create comment editors, but they
use a light shade of gray and skip the underline decoration, making
them less visually distracting.

Skip lines now use a paler shade of blue and also hide the fact they
contain anchors, until you hover over them and the anchor shows up.

The expand before and after are changed to be arrows showing in
which direction the lines will appear above or below the skip
line. This makes it easier for people like myself who get confused
and always hit the wrong link, the arrow shows to which side the
lines will appear.

Change-Id: I7108733358db23dca6f7d5b90cd4b7e8712cf2b2

Former-commit-id: 412150cf8b3190b575cc426279a1a22e1e562a8a