Fixed cleanup of propagated SshScopes
The RequestScopePropagator handles RequestCleanup independently of
the original context. However, SshScope.Context.subContext() was
chaining the cleanup in the parent. This would cause an exception
if the parent scope happened to finish before the sub-context. Since
cleanup is already handled correctly by the RequestScopePropagator,
don't register the continued contect with the original context cleanup.

Change-Id: Ie508d1921050887b95dd18bc5806c17e4894391f

Former-commit-id: 2571ceb388fb885fd1189c653c9aab6845e88cc8