Fix case check for project name so that symlinks work again
The check for the case of the project name when opening a git
repository that was introduced by change 30015 (commit
54217d49bf2f20994f53af247493bebcbc76df7a [formerly a7e928daaf7ba4176b5044797f90b21f420ab2f7], "Verify the case of the
project name before opening git repository") prevents the usage of
symbolic links under the git base path. There are some scenarios
where symbolic links for git repositories make sense. E.g. if a
project was renamed it may make sense to create a symbolic link
with the old name that points to the new location so that old
references to this project can still be resolved.

The case check for the project name is now implemented in such a
way that it compares the input project name against a cached list of
all existing project names. With this new implementation symbolic
links for git repositories should work again.

Bug: issue 1353
Change-Id: Iad2be143fd7e2057700cc447f3426df732f9b8cb
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 9dbe37a8a30908a539b3fe705b4f95db9fadcd65