Delete unused onSignOut method
Once upon a time Gerrit performed a sign-out by just deleting the
cookie and updating the DOM to remove signed-in content. This is
not true anymore, the entire page is refreshed to delete any data
that is private to a signed-in account, and to ensure the server
can clear the correct HttpOnly cookies to actually do a sign-out.

Since that behavior change, this onSignOut method is dead code.

Change-Id: I7d7922f3a38fdcd03615866d0d693088660e1a31

Former-commit-id: da8fee1694d57494320246b02580dd6befb7fa22