Cache IdentifiedUser across PrologEnvironments
When running multiple rules for the same change, the set of
reviewers is identical as ChangeControl goes up the project
tree and runs rules in parent projects. Recycle all of the
IdentifiedUser objects that were made in one invocation for
the next one, to avoid looking them up again in AccountCache.

Change-Id: I4031a3f8f9eb0328f0abbbfce1f38eb67b4877d3

Former-commit-id: 4502dfbde0c12b360e11e3a46ee3df13b2ea691c