Define gerrit-plugin-api as a subset API
This is a new JAR produced by the build that contains almost all of
the code that the corresponding WAR contains at runtime. This permits
plugin authors to compile against this single interface JAR, and then
load their plugin into a running server that uses the real WAR.

Having the API be a single JAR simplifies plugin development, Maven
users can depend on this single JAR using the provided scope. Other
IDE users can just make a new Java project and add this single JAR
as the only dependency.

Change-Id: Ie045ec4202d44b78b75ccb4af000e13c1e7378ce

Former-commit-id: 980efdb3768ab063bf641fabad8a1214c3954e82